// src/services/PregnancyRegisterService/interface.ts
export interface PregnancyRegisterInterface {
  dataUltimaMenstruacao: string | string[] | undefined;
  gestanteId: number | undefined;
  dataInicioGestacao: string | string[] | undefined;
  fatorRh: string | undefined;
  frequenciaUsoAlcool: number | undefined;
  fuma: boolean | undefined;
  gravidezPlanejada: boolean | undefined;
  grupoSanguineo: number | undefined;
  pesoAntesGestacao: number | undefined;
  quantidadeCigarrosDia: number | undefined;
  riscoGestacional: number | undefined;
  situacaoGestacional: number | undefined;
  usoDrogas: number | undefined;
  vacinaHepatiteB: boolean | undefined;
  consumoAlcool: boolean | undefined;
}